refactor research network components to be compatible with both phase 1 and 2
